fix: compute the exact binomial coefficient with integer arithmetic
pairwiseMargin's exact binomial ran in floating point in Python, TypeScript,
Kotlin, and R. For C(56,27) the partial products (result*(n-i) before the
divide) exceed 2^53, so float rounding accumulated and the margin came out
784 instead of the correct 782 that Go, C#, and Rust produce with integer
arithmetic. Switch Python/TS/Kotlin to integer partial products (bigint,
BigInt, Long), casting to double only at the end.

R has no exact integer type wide enough, and its built-in choose() is
inexact even below 2^53 (choose(56,27) returns ...078, not the exact ...080).
Give R a native 64-bit helper (src/binomial.c) mirroring the Go/Rust
algorithm, and route both the exact pairwise branch and
min_achievable_misrate_two_sample through it, so all seven ports agree
bit-for-bit.

The downstream half-margin clamp masked this in bounds output, but the raw
margins disagreed across languages -- a trap for future fixture regeneration.
Regression tests pin the value at 782.

Acked-by: Andrey Akinshin <andrey.akinshin@gmail.com>

+++ kt/src/main/kotlin/dev/pragmastat/PairwiseMargin.kt
@@ -248,13 +248,13 @@ private fun binomialCoefficient(
     if (kk == 0 || kk == n) return 1.0
 
     kk = minOf(kk, n - kk) // Take advantage of symmetry
-    var result = 1.0
+    var result = 1L // exact integer arithmetic: each partial product is divisible
 
     for (i in 0 until kk) {
         result = result * (n - i) / (i + 1)
     }
 
-    return result
+    return result.toDouble()
 }
 
 /**
